refactor: move EventBus from game to common utils
